move ui-type to respective folder
