+++ src/app/pipes/to-grid-items/to-grid-items.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { MediaToGridMediaService } from 'src/app/services/media-to-grid-media.service';
-import { GridMediaItem } from 'src/app/types/grid-media-item';
+import { GridMediaItem } from 'src/app/types/ui-types/grid-media-item';
 import { MediaItem } from 'src/app/types/media-item';
 
 @Pipe({
